refactor searchbar component using BRC
