+++ src/components/Base/Dropdown/dropdown.scss
@@ -1,4 +1,4 @@
-@import '../../theme/variables.scss';
+@import '../../../theme/variables.scss';
 
 $border-color-open: $azure;
