+++ src/components/Base/Input/Input.tsx
@@ -1,0 +1,3 @@
+import { Input } from 'buildo-react-components/lib/Input/Input';
+import './input.scss';
+export default Input;
