+++ src/components/Base/Input/input.scss
@@ -1,0 +1,1 @@
+@import '~buildo-react-components/src/Input/input.scss';
