+++ src/components/Base/StatefulButton/StatefulButton.tsx
@@ -1,0 +1,3 @@
+import { StatefulButton } from 'buildo-react-components/lib/Button';
+import './statefulButton.scss';
+export default StatefulButton;
