+++ src/components/Base/StatefulButton/statefulButton.scss
@@ -1,0 +1,1 @@
+@import '~buildo-react-components/src/Button/button.scss';
